requirements/dev.txt
-r base.txt
black==23.10.1
isort==5.12.0
flake8-django==1.4